R code to create Shiny App for Simple Dashboard
This is R code to create an interactive Shiny UI dashboard that displays descriptive statistics (basics), along with choice of histogram, boxplot, violin, or density plots that can be scrolled over. The server logic allows grouping and subsetting (on categorical variables and a single continuous variable range). This also offers the options to download a data table with descriptive statistics and an entry for returning percentile ranks for given values on outcome variable.